+++ ScaleCalculator/ScaleCalculator/Measures.h
@@ -11,7 +11,7 @@ struct MeasuredDirection;
 struct Point {
 	double x{0}, y{0};
 	Point() = default;
-	Point(int x, int y);
+	Point(double x, double y);
 };
 
 /// <summary>
